separated logic from presentation in sale search list
