+++ src/components/ApplicationViews.js
@@ -4,7 +4,7 @@ import DashBoard from "../components/dashboard/DashBoard.js"
 import DealershipList from "./dealership/controllers/searchListContainer"
 import EmployeeList from "./employee/controllers/searchListContainer"
 import VehiclesList from "./vehicle/controllers/searchListContainer"
-import SaleList from "./sale/list"
+import SaleList from "./sale/searchListContainer"
 
 
 const ApplicationViews = props => {
